feat: Initialize game with deck
When running main, if a file is provided as an argument, it will read a
comma separated list of cards from the file. A game will be created with
these cards as the deck. If no arguments are provided, it will just
instanciate a new deck (which currently will create a deck of all the
cards in random order).